third_party/blink/web_tests/fast/xpath/4XPath/Core/test_boolean_expr-expected.txt

PASS doc.evaluate("3.14 < -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("3.14 <= -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("3.14 > -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("3.14 >= 3.14", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("number('NaN') < -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("number('NaN') <= -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("number('NaN') > -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("number('NaN') >= 3.14", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/@num2 <= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/root/@num2 >= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/root/@num0 <= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/root/@num0 >= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/@num4 <= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/@num4 >= /root/@num2", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("'3.14' = '3.14'", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("'3.14' = 'Hi'", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("'3.14' != -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("number('NaN') = 'Hi'", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("number('NaN') != -4", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("number('NaN')  = number('NaN')", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("number('NaN') != number('NaN')", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/empty  = true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/empty != true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/empty  = false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/empty != false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/empty  = /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/empty != /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/*[1]  = /empty", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/*[1] != /empty", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("/root/*[1]  = /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("/root/*[1] != /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("'egg1'  = /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("'egg1' != /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("'egg2'  = /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("'egg2' != /root/*[1]", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("'egg1'  = /root/*", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("'egg1' != /root/*", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("true() and true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("true() and false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("false() and true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("false() and false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS doc.evaluate("true() or false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("false() or true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("true() or true()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is true
PASS doc.evaluate("false() or false()", doc, null, XPathResult.BOOLEAN_TYPE, null).booleanValue is false
PASS successfullyParsed is true

TEST COMPLETE
